API-solutely fabulous - routes
