+++ client/index.tsx
@@ -1,7 +1,11 @@
 import { createRoot } from 'react-dom/client'
+import { RouterProvider } from 'react-router-dom'
 
-import App from './components/App'
+import { router } from './router'
+// import AppLayout from './components/AppLayout'
 
 document.addEventListener('DOMContentLoaded', () => {
-  createRoot(document.getElementById('app') as HTMLElement).render(<App />)
+  createRoot(document.getElementById('app') as HTMLElement).render(
+    <RouterProvider router={router} />
+  )
 })
