+++ package.json
@@ -28,7 +28,8 @@
   },
   "dependencies": {
     "cors": "^2.8.5",
-    "express": "^4.17.2"
+    "express": "^4.17.2",
+    "react-router-dom": "^6.12.0"
   },
   "devDependencies": {
     "@devacademy/eslint-config": "^1.8.0",
